list conversion error prelim analysis

+++ Statistics.Test/StatsUnitTest.cs
@@ -11,7 +11,8 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<double>{1.5, 8.9, 3.2, 4.5});
+                new List<float>
+                {(float)1.5,(float) 8.9, (float)3.2,(float) 4.5});
             float epsilon = 0.001F;
             Assert.True(Math.Abs(computedStats.average - 4.525) <= epsilon);
             Assert.True(Math.Abs(computedStats.max - 8.9) <= epsilon);
@@ -22,7 +23,7 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<double>{});
+                new List<float>{});
 
             Assert.True(Double.IsNaN(computedStats.average));
             Assert.True(Double.IsNaN(computedStats.max));
